EES-5531 Move `AppOptions` to `Options` namespace in other subprojects
