+++ src/GovUk.Education.ExploreEducationStatistics.Notifier/Services/ApiSubscriptionTableStorageService.cs
@@ -1,5 +1,5 @@
 using GovUk.Education.ExploreEducationStatistics.Common.Services;
-using GovUk.Education.ExploreEducationStatistics.Notifier.Configuration;
+using GovUk.Education.ExploreEducationStatistics.Notifier.Options;
 using GovUk.Education.ExploreEducationStatistics.Notifier.Services.Interfaces;
 using Microsoft.Extensions.Options;
